+++ src/lib/bunker/index.ts
@@ -4,6 +4,7 @@
 
 export {
   isRunning,
+  restartBunkerSubscriptions,
   startBunker,
   stopAllBunkers,
   stopBunker,